fixed signup validation form // added some styling to login form // merged custom useForm hooks
