+++ src/App.scss
@@ -4,10 +4,11 @@
 @import './styles/sass/base/reset';
 
 @import './styles/sass/pages/places';
+@import './styles/sass/pages/signupForm';
+@import './styles/sass/pages/loginForm';
 @import './styles/sass/pages/bookings';
 @import './styles/sass/layout/header';
 @import './styles/sass/layout/footer';
 @import './styles/sass/layout/nav';
-@import './styles/sass/pages/signupForm';
 @import './styles/sass/layout/forms';
 @import './styles/sass/layout/topnav';
